Split into interface and implementation
